+++ src/ldpred2-auto.cpp
@@ -6,6 +6,9 @@
 
 /******************************************************************************/
 
+const double MIN_P  = 1e-5;
+const double MIN_H2 = 1e-4;
+
 inline double square(double x) {
   return x * x;
 }
@@ -34,7 +37,6 @@ List ldpred2_gibbs_auto(Environment corr,
   myassert_size(beta_init.size(), m);
   myassert_size(n_vec.size(), m);
 
-  std::vector<double> is_causal(m, p_init);
   arma::vec curr_beta(beta_init.begin(), m);
   arma::vec avg_beta(m, arma::fill::zeros), avg_postp(m, arma::fill::zeros);
 
@@ -44,42 +46,46 @@ List ldpred2_gibbs_auto(Environment corr,
   int num_iter_tot = burn_in + num_iter;
   std::vector<double> p_est(num_iter_tot), h2_est(num_iter_tot);
 
-  double nb_causal = m * p_init;
   double cur_h2_est = arma::dot(curr_beta, sfbm->prod(curr_beta));
   double p = p_init, h2 = h2_init, avg_p = 0, avg_h2 = 0;
 
   for (int k = 0; k < num_iter_tot; k++) {
+
+    int nb_causal = 0;
+    double h2_per_var = h2 / (m * p);
+    double inv_odd_p = (1 - p) / p;
 
     for (const int& j : order) {
 
       double dotprod = sfbm->dot_col(j, curr_beta);
       double res_beta_hat_j = beta_hat[j] + curr_beta[j] - dotprod;
 
-      double C1 = h2 * n_vec[j] / (m * p);
+      double C1 = h2_per_var * n_vec[j];
       double C2 = 1 / (1 + 1 / C1);
-      double C3 = C2 * res_beta_hat_j ;
+      double C3 = C2 * res_beta_hat_j;
       double C4 = ::sqrt(C2 / n_vec[j]);
 
       double postp = 1 /
-        (1 + (1 - p) / p * ::sqrt(1 + C1) * ::exp(-square(C3 / C4) / 2));
+        (1 + inv_odd_p * ::sqrt(1 + C1) * ::exp(-square(C3 / C4) / 2));
 
       if (k >= burn_in) {
         avg_postp[j] += postp;
         avg_beta[j]  += C3 * postp;
       }
 
-      double nb_rm = is_causal[j];
-      is_causal[j] = postp > ::unif_rand();
-      nb_causal += is_causal[j] - nb_rm;
-
-      double prev_beta = curr_beta[j];
-      curr_beta[j] = is_causal[j] ? (C3 + ::norm_rand() * C4) : 0;
-      double diff = curr_beta[j] - prev_beta;
+      double diff = -curr_beta[j];
+      if (postp > ::unif_rand()) {
+        curr_beta[j] = ::Rf_rnorm(C3, C4);
+        diff += curr_beta[j];
+        nb_causal++;
+      } else {
+        curr_beta[j] = 0;
+      }
       cur_h2_est += diff * (2 * dotprod + diff);
     }
 
-    p = ::Rf_rbeta(1 + nb_causal, 1 + m - nb_causal);
-    h2 = std::max(cur_h2_est, 1e-4);
+    p = std::max(::Rf_rbeta(1 + nb_causal, 1 + m - nb_causal), MIN_P);
+    h2 = std::max(cur_h2_est, MIN_H2);
     if (verbose) Rcout << k + 1 << ": " << p << " // " << h2 << std::endl;
 
     if (k >= burn_in) {