+++ tests/testthat/test-4-pcadapt.R
@@ -13,7 +13,6 @@ expect_s4_class(G, "FBM.code256")
 
 test_that("Same as pcadapt", {
 
-  # skip_on_cran()
   skip_if_not_installed("pcadapt")
 
   ################################################################################